app/src/main/java/com/example/coursescheduler/objects/Student.java
package com.example.coursescheduler.objects;

import java.util.Objects;

public class Student {

    private final String studentID;
    private final String studentName;

    public Student(final String studentID, final String studentName){
        this.studentID = studentID;
        this.studentName = studentName;
    }

    public String getStudentID() {
        return studentID;
    }

    public String getStudentName() {
        return studentName;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Student student = (Student) o;
        return Objects.equals(studentID, student.studentID) &&
                Objects.equals(studentName, student.studentName);
    }

    @Override
    public int hashCode() {
        return Objects.hash(studentID, studentName);
    }
}
